requirements.txt
discord.py==1.7.3
psycopg-binary==3.0.5
python-dotenv==0.19.2
